extracted a ton of code into individual files under a new dir 'source/js/foundation
