+++ source/js/foundation/index.js
@@ -1,0 +1,2 @@
+export { bindEventHandlers } from "./template-js-handler";
+export { injectReactComponents } from "./inject-react";
