+++ source/js/foundation/inject-react/index.js
@@ -1,0 +1,18 @@
+import injectMultipageNav from "./multipage-nav.js";
+import injectNews from "./news.js";
+import injectPetition from "./petition.js";
+import injectPulseProjectList from "./pulse-project-list.js";
+import injectShareButtonGroup from "./share-button-group.js";
+
+/**
+ * Inject React components
+ * @param {Array} apps The existing array we are using to to track all ReactDOM.render calls
+ * @param {String} siteUrl Foundation site base URL
+ */
+export const injectReactComponents = (apps, siteUrl, env) => {
+  injectMultipageNav(apps);
+  injectNews(apps, env);
+  injectPetition(apps, siteUrl);
+  injectPulseProjectList(apps, env);
+  injectShareButtonGroup(app);
+};
