+++ source/js/foundation/template-js-handler/global/sticky-cta.js
@@ -1,0 +1,7 @@
+/**
+ * Bind handler to
+ *
+ */
+export default () => {
+  // to be added
+};
