+++ source/js/foundation/template-js-handler/pulse-profile-list-filter.js
@@ -1,0 +1,22 @@
+import { ReactGA } from "../../common/react-ga-proxy.js";
+
+/**
+ * Bind click handler to
+ * ".profile-directory .fellowships-directory-filter .filter-option button"
+ */
+export default () => {
+  document
+    .querySelectorAll(
+      `.profile-directory .fellowships-directory-filter .filter-option button`
+    )
+    .forEach(filter => {
+      let year = filter.textContent.trim();
+      filter.addEventListener(`click`, () => {
+        ReactGA.event({
+          category: `profiles`,
+          action: `directory filter`,
+          label: `${document.title} ${year}`
+        });
+      });
+    });
+};
