+++ source/js/main.js
@@ -11,11 +11,7 @@ import {
   ReactGA
 } from "./common";
 
-import Petition from "./components/petition/petition.jsx";
-import MultipageNavMobile from "./components/multipage-nav-mobile/multipage-nav-mobile.jsx";
-import News from "./components/news/news.jsx";
-import PulseProjectList from "./components/pulse-project-list/pulse-project-list.jsx";
-import ShareButtonGroup from "./components/share-button-group/share-button-group.jsx";
+import { bindEventHandlers, injectReactComponents } from "./foundation";
 
 import primaryNav from "./primary-nav.js";
 import bindMozFestGA from "./mozfest-ga.js";
@@ -224,39 +220,6 @@ let main = {
   },
 
   bindGAEventTrackers() {
-    let seeMorePage = document.querySelector(`#see-more-modular-page`);
-
-    if (seeMorePage) {
-      let label = ``;
-      let pageHeader = document.querySelector(`.cms h1`);
-
-      if (pageHeader) {
-        label = `${pageHeader.innerText} - footer cta`;
-      }
-
-      seeMorePage.addEventListener(`click`, () => {
-        ReactGA.event({
-          category: `navigation`,
-          action: `page footer cta`,
-          label: label
-        });
-      });
-    }
-
-    let participateDonateBtn = document.querySelector(
-      `#view-participate .card-cta .btn[href*="donate.mozilla.org"]`
-    );
-
-    if (participateDonateBtn) {
-      participateDonateBtn.addEventListener(`click`, () => {
-        ReactGA.event({
-          category: `donate`,
-          action: `donate button tap`,
-          label: document.title
-        });
-      });
-    }
-
     // MozFest specific scripts
     bindMozFestGA();
     bindMozFestEventHandlers();
@@ -265,248 +228,9 @@ let main = {
   // Embed various React components based on the existence of containers within the current page
   injectReactComponents() {
     injectCommonReactComponents(apps, networkSiteURL, csrfToken);
+    injectReactComponents(apps, networkSiteURL, env);
 
-    // petition elements
-    var subscribed = false;
-
-    if (window.location.search.indexOf(`subscribed=1`) !== -1) {
-      subscribed = true;
-    }
-
-    document.querySelectorAll(`.sign-petition`).forEach(element => {
-      var props = element.dataset;
-
-      props.apiUrl = `${networkSiteURL}/api/campaign/petitions/${props.petitionId}/`;
-
-      apps.push(
-        new Promise(resolve => {
-          ReactDOM.render(
-            <Petition
-              {...props}
-              isHidden={false}
-              subscribed={subscribed}
-              whenLoaded={() => resolve()}
-            />,
-            element
-          );
-        })
-      );
-    });
-
-    // Multipage nav used in landing pages
-    if (document.querySelector(`#multipage-nav`)) {
-      let links = [];
-
-      links = [].map.call(
-        document.querySelectorAll(`#multipage-nav a`),
-        child => {
-          return {
-            label: child.textContent.trim(),
-            href: child.getAttribute(`href`),
-            isActive: !!child.getAttribute(`class`).match(/active/)
-          };
-        }
-      );
-
-      apps.push(
-        new Promise(resolve => {
-          ReactDOM.render(
-            <MultipageNavMobile links={links} whenLoaded={() => resolve()} />,
-            document.querySelector(
-              `#multipage-nav-mobile .container .row .col-12`
-            )
-          );
-        })
-      );
-    }
-
-    // News
-    if (document.querySelector(`#news`)) {
-      apps.push(
-        new Promise(resolve => {
-          ReactDOM.render(
-            <News env={env} whenLoaded={() => resolve()} />,
-            document.querySelector(`#news`)
-          );
-        })
-      );
-    }
-
-    // Pulse project lists
-    document.querySelectorAll(`.pulse-project-list`).forEach(target => {
-      apps.push(
-        new Promise(resolve => {
-          ReactDOM.render(
-            <PulseProjectList
-              env={env}
-              featured={target.dataset.featured === `True`}
-              help={target.dataset.help}
-              issues={target.dataset.issues}
-              max={parseInt(target.dataset.max, 10)}
-              query={target.dataset.query || ``}
-              reverseChronological={target.dataset.reversed === `True`}
-              whenLoaded={() => resolve()}
-              directLink={target.dataset.directLink === `True`}
-            />,
-            target
-          );
-        })
-      );
-    });
-
-    // Share button group
-    document
-      .querySelectorAll(`.share-button-group-wrapper`)
-      .forEach(element => {
-        var props = element.dataset;
-
-        apps.push(
-          new Promise(resolve => {
-            ReactDOM.render(
-              <ShareButtonGroup {...props} whenLoaded={() => resolve()} />,
-              element
-            );
-          })
-        );
-      });
-
-    //Profile Directory Filter-Bar GA
-
-    document
-      .querySelectorAll(
-        `.profile-directory .fellowships-directory-filter .filter-option button`
-      )
-      .forEach(filter => {
-        let year = filter.textContent.trim();
-        filter.addEventListener(`click`, () => {
-          ReactGA.event({
-            category: `profiles`,
-            action: `directory filter`,
-            label: `${document.title} ${year}`
-          });
-        });
-      });
-
-    //Profile Directory Cards Social Media GA
-
-    function profileCardSocialAnalytics(
-      socialTwitter,
-      socialLinkedIn,
-      profileName
-    ) {
-      if (socialTwitter) {
-        socialTwitter.addEventListener(`click`, () => {
-          ReactGA.event({
-            category: `profiles`,
-            action: `profile tap`,
-            label: `${document.title} ${profileName} twitter`,
-            transport: `beacon`
-          });
-        });
-      }
-
-      if (socialLinkedIn) {
-        socialLinkedIn.addEventListener(`click`, () => {
-          ReactGA.event({
-            category: `profiles`,
-            action: `profile tap`,
-            label: `${document.title} ${profileName} linkedin`,
-            transport: `beacon`
-          });
-        });
-      }
-    }
-
-    //Profile Directory Card Headshot/Name GA
-
-    function bindProfileCardAnalytics(profileCards) {
-      // event listener & GA
-      let bindAnalytics = (element, profileName) => {
-        element.addEventListener(`click`, () => {
-          ReactGA.event({
-            category: `profiles`,
-            action: `profile tap`,
-            label: `${document.title} ${profileName} pulse profile`,
-            transport: `beacon`
-          });
-        });
-      };
-
-      // adding event listener for each headshot & name
-      profileCards.forEach(card => {
-        let profileHeadshotElement = card.querySelector(`.headshot-container`);
-        let profileNameElement = card.querySelector(`.meta-block-name`);
-        let profileName = profileNameElement.textContent.trim();
-
-        [(profileNameElement, profileHeadshotElement)].forEach(target =>
-          bindAnalytics(target, profileName)
-        );
-
-        let socialTwitter = card.querySelector(`.twitter`);
-        let socialLinkedIn = card.querySelector(`.linkedIn`);
-        profileCardSocialAnalytics(socialTwitter, socialLinkedIn, profileName);
-      });
-    }
-
-    // store profile cards
-    function updateProfileList() {
-      let profileCards = document.querySelectorAll(`.profiles .person-card`);
-      bindProfileCardAnalytics(profileCards);
-    }
-
-    // Checks for profile cards in the initial page load
-    updateProfileList();
-
-    // And start listening for profile filter events,
-    // in case profile cards get updated.
-    document.addEventListener(`profiles:list-updated`, () =>
-      updateProfileList()
-    );
-
-    // Enable the "load more results" button on index pages
-    let loadMoreButton = document.querySelector(`.load-more-index-entries`);
-    if (loadMoreButton) {
-      const entries = document.querySelector(`.index-entries`);
-
-      // Get the page size from the document, which the IndexPage should
-      // have templated into its button as a data-page-size attribute.
-      const pageSize = parseInt(loadMoreButton.dataset.pageSize) || 12;
-
-      // Start at page 1, as page 0 is the same sat as the initial page set.
-      let page = 1;
-
-      const loadMoreResults = () => {
-        loadMoreButton.disabled = true;
-
-        // Construct our API call as a relative URL:
-        let url = `./entries/?page=${page++}&page_size=${pageSize}`;
-
-        // And then fetch the results and render them into the page.
-        fetch(url)
-          .then(result => result.json())
-          .then(data => {
-            if (!data.has_next) {
-              loadMoreButton.removeEventListener(`click`, loadMoreResults);
-              loadMoreButton.parentNode.removeChild(loadMoreButton);
-            }
-            return data.entries_html;
-          })
-          .then(entries_html => {
-            const div = document.createElement(`div`);
-            div.innerHTML = entries_html;
-            Array.from(div.children).forEach(c => entries.appendChild(c));
-          })
-          .catch(err => {
-            // TODO: what do we want to do in this case?
-            console.error(err);
-          })
-          .finally(() => {
-            loadMoreButton.disabled = false;
-          });
-      };
-
-      loadMoreButton.addEventListener(`click`, loadMoreResults);
-    }
+    bindEventHandlers();
   }
 };
